ConnectToDatabase.java (Configuration Bean) is now responsible for taking the arguments providing it to the DatabaseService.
DatabaseService.connectToDatabase tries to connect to the database and throws exceptions if that fails instead of just providing a status message. That means that the complete Application executes (because why would want it to keep on running if there is no connection to a database?). That same method also tests if the URL it used also contained the databases' name. If not, an exception is thrown.

The RESTAPITest requires arguments (see RESTAPITest.java).

README.adoc is updated.