Dependency Handling
One of the dependencies declared in the pom.xml provided a transitive dependency to jakarta.xml.bind:jakarta.xml.bind-api:4.0.1, which had a package named jakarta.xml.bind. But the XEStoWorkflowConverter depended on classes from javax.xml.bind, which is not in bind-api:4.0.1, but in version 2.3.3. Changing the pom.xml accordingly switched the versions (there is only one jakarta.xml.bind:jakarta.xml.bind-api version!) and thus solved