Adjusted pom.xml (jakarta.xml.bind)
So the problem was that the XEStoWorkflowConverter depended on version 2.3.3 of the jakarta.xml.bind-api, but the restapi depended on version 4.0.1. That is why I changed the version in the restapi in one of the last commits, not realizing that that would create another conflict. The XEStoWorkflowConverter now simply depends on version 4.0.1, as well as the restapi. Let's hope this did not create further conflicts...